IPNetworkField model field uses like form field
The IPNetworkField model field now uses the IPNetworkField form field by
default when rendered in a ModelForm.

In making this change, inheritance of IPNetworkField (model field) was
changed to inherit from Field instead of CharField, as the former
structure was causing issues when calling the formfield() method;
however, this then resulted in Django not knowing how to create the
field in the database until adding the get_internal_type() method, which
now returns 'CharField'. The net effect is that, roughly speaking,
database-side the field inherits from CharField, but otherwise it
inherits directly from Field instead.