Rename to CIDRNetworkField
Renamed the model field IPNetworkField to CIDRNetworkField, as this field does in fact store a CIDR network. This also avoids confusion with the IPNetworkField form field, which accepts netmask-style networks in addition to CIDR-style. Both fields also accept individual IP addresses; this is reasoned away as the assumption of a default /32 network when not otherwise specified.